chore(cli): further refactor kamel local cmds
